+++ lib/src/typedefs.dart
@@ -12,14 +12,14 @@ typedef JSON = Map<String, dynamic>;
 /// `String` to `Map<String, dynamic>` documents.
 typedef JsonCollection = Map<String, Map<String, dynamic>>;
 
-/// An alias for `int`, used to denote the frequency of a [Term] in an index or
+/// An alias for `int`, used to denote the frequency of a term in an index or
 /// indexed object.
 typedef Ft = int;
 
-/// Alias for `Map<Zone, double>`.
+/// Alias for `Map<String, double>`.
 ///
 /// Maps the zone names to their relative weight.
-typedef ZoneWeightMap = Map<Zone, double>;
+typedef ZoneWeightMap = Map<String, double>;
 
 /// Alias for `double` where it represents the inverse document frequency of a
 /// term.
@@ -36,7 +36,7 @@ typedef IdFt = double;
 ///
 /// Maps the vocabulary to hashmaps of document id to term frequency in the
 /// document.
-typedef FtdPostings = Map<Term, Map<DocId, Ft>>;
+typedef FtdPostings = Map<String, Map<String, Ft>>;
 
 /// Alias for `Map<String, int>`.
 ///
@@ -46,12 +46,12 @@ typedef FtdPostings = Map<Term, Map<DocId, Ft>>;
 /// key and the document frequency as the values.
 ///
 /// A [DftMap] can be an asynchronous data source or an in-memory hashmap.
-typedef DftMap = Map<Term, Ft>;
+typedef DftMap = Map<String, Ft>;
 
-/// Alias for `MapEntry<Term, Ft>`.
+/// Alias for `MapEntry<String, Ft>`.
 ///
 /// A [DftMapEntry] is an entry in a [DftMap].
-typedef DftMapEntry = MapEntry<Term, Ft>;
+typedef DftMapEntry = MapEntry<String, Ft>;
 
 /// A callback that asynchronously retrieves a [DftMap] subset for a collection
 /// of [terms] from a [DftMap] data source, usually persisted storage.
@@ -72,15 +72,15 @@ typedef CollectionSizeCallback = Future<int> Function();
 
 /// Alias for `Map<String, double>`.
 ///
-/// Maps the vocabulary [Term] to [IdFt].
-typedef IdFtIndex = Map<Term, IdFt>;
+/// Maps the vocabulary term to [IdFt].
+typedef IdFtIndex = Map<String, IdFt>;
 
 /// A callback that asynchronously retrieves a [KGramsMap] subset for a
 /// collection of [terms] from a [KGramsMap] data source,
 /// usually persisted storage.
 ///
 /// Loads the entire [KGramsMap] if [terms] is null.
-typedef KGramsMapLoader = Future<KGramsMap> Function([Iterable<Term>? terms]);
+typedef KGramsMapLoader = Future<KGramsMap> Function([Iterable<String>? terms]);
 
 /// A callback that passes [values] for persisting to a [KGramsMap].
 ///
@@ -88,57 +88,54 @@ typedef KGramsMapLoader = Future<KGramsMap> Function([Iterable<Term>? terms]);
 /// entries.
 typedef KGramsMapUpdater = Future<void> Function(KGramsMap values);
 
-/// Type definition for a hashmap of [Term] to [DocPostingsMap].
-typedef PostingsMap = Map<Term, DocPostingsMap>;
+/// Type definition for a hashmap of term to [DocPostingsMap].
+typedef PostingsMap = Map<String, DocPostingsMap>;
 
-/// Type definition for a hashmap entry of [Term] to [DocPostingsMap] in a
+/// Type definition for a hashmap entry of term to [DocPostingsMap] in a
 /// [PostingsMap] hashmap.
 ///
 /// Alias for `MapEntry<String, Map<String, Map<String, List<int>>>>`.
-typedef PostingsMapEntry = MapEntry<Term, DocPostingsMap>;
+typedef PostingsMapEntry = MapEntry<String, DocPostingsMap>;
 
 /// Type definition for a hashmap of keywords to a map of document ids to
 /// keyword scores for that keyword in each document.
 ///
 /// Alias for `Map<String, Map<String, double>>`.
-typedef KeywordPostingsMap = Map<Term, KeyWordPostings>;
+typedef KeywordPostingsMap = Map<String, KeyWordPostings>;
 
 /// Type definition for a hashmap of document ids to keyword scores for a
 /// keyword in each document.
 ///
 /// Alias for `Map<String, double>`.
-typedef KeyWordPostings = Map<DocId, double>;
+typedef KeyWordPostings = Map<String, double>;
 
-/// An alias for [int], used to denote the position of a [Term] in [SourceText]
+/// An alias for [int], used to denote the position of a term in [String]
 /// indexed object (the term position).
 ///
 /// Alias for `int`;
 typedef Pt = int;
 
-/// An alias for [String], used whenever a document id is referenced.
-typedef DocId = String;
-
-/// Type definition for a hashmap of [DocId] to [ZonePostingsMap].
+/// Type definition for a hashmap of document id to [ZonePostingsMap].
 ///
 /// Alias for `Map<String, List<int>>`.
-typedef DocPostingsMap = Map<DocId, ZonePostingsMap>;
+typedef DocPostingsMap = Map<String, ZonePostingsMap>;
 
-/// Type definition for a hashmap entry of [DocId] to [ZonePostingsMap] in a
+/// Type definition for a hashmap entry of document id to [ZonePostingsMap] in a
 /// [DocPostingsMap] hashmap.
 ///
 /// Alias for `MapEntry<String, Map<String, List<int>>>`.
-typedef DocPostingsMapEntry = MapEntry<DocId, ZonePostingsMap>;
+typedef DocPostingsMapEntry = MapEntry<String, ZonePostingsMap>;
 
-/// Type definition for a hashmap of [Zone]s to [TermPositions].
+/// Type definition for a hashmap of zones to [TermPositions].
 ///
 /// Alias for `Map<String, List<int>>`.
-typedef ZonePostingsMap = Map<Zone, TermPositions>;
+typedef ZonePostingsMap = Map<String, TermPositions>;
 
-/// Type definition for a hashmap entry of [Zone] to [TermPositions] in a
+/// Type definition for a hashmap entry of zone to [TermPositions] in a
 /// [ZonePostingsMap] hashmap.
 ///
 /// Alias for `MapEntry<String, List<int>>`.
-typedef ZonePostingsMapEntry = MapEntry<Zone, TermPositions>;
+typedef ZonePostingsMapEntry = MapEntry<String, TermPositions>;
 
 /// Type definition for an ordered [Set] of unique zero-based term
 /// positions in a text source, sorted in ascending order.
@@ -146,7 +143,8 @@ typedef TermPositions = List<Pt>;
 
 /// Asynchronously retrieves a [PostingsMap] subset for a collection of
 /// [terms] from a [PostingsMap] data source, usually persisted storage.
-typedef PostingsMapLoader = Future<PostingsMap> Function(Iterable<Term> terms);
+typedef PostingsMapLoader = Future<PostingsMap> Function(
+    Iterable<String> terms);
 
 /// A callback that [values] for persisting to a [PostingsMap].
 ///
@@ -157,7 +155,7 @@ typedef PostingsMapUpdater = Future<void> Function(PostingsMap values);
 /// Asynchronously retrieves a [KeywordPostingsMap] subset for a collection of
 /// [keywords] from a [KeywordPostingsMap] data source, usually persisted storage.
 typedef KeywordPostingsMapLoader = Future<KeywordPostingsMap> Function(
-    Iterable<Term> keywords);
+    Iterable<String> keywords);
 
 /// A callback that [values] for persisting to a [KeywordPostingsMap].
 ///